+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
  - Rename package to `zarrs_ffi` from `zarrs-ffi` and move repository
  - `zarrsDestroyArray` and `zarrsDestroyStorage` now return a `ZarrsResult`
+ - Set MSRV to 1.71
 
 ## [0.4.0] - 2024-03-09